Basic PartiqlAst.Expr converter
- Places rel and rex together in plan.ir
- Adds debug to shell
- Adds basic plan pretty printer
- Need to follow up on Jackson Map<T,S> serde
