+++ partiql-cli/src/main/kotlin/org/partiql/cli/Main.kt
@@ -13,13 +13,16 @@
  */
 @file:JvmName("Main")
 
-@file:Suppress("DEPRECATION")
-
 package org.partiql.cli
 
 import com.amazon.ion.system.IonSystemBuilder
 import org.partiql.cli.pico.PartiQLCommand
+import org.partiql.lang.eval.EvaluationSession
+import org.partiql.lang.syntax.PartiQLParserBuilder
+import org.partiql.plan.AstToRel
+import org.partiql.plan.debug.PlanPrinter
 import picocli.CommandLine
+import java.io.PrintStream
 import kotlin.system.exitProcess
 
 /**
@@ -30,4 +33,23 @@ fun main(args: Array<String>) {
     val command = CommandLine(PartiQLCommand(ion))
     val exitCode = command.execute(*args)
     exitProcess(exitCode)
+}
+
+/**
+ * Highly visible place to modify shell behavior for debugging
+ *
+ * Consider giving this access to the print stream in Shell.
+ * It would have been too hacky without a slight refactor, so now let's just assume System.out for debugging
+ */
+object Debug {
+
+    @Throws(Exception::class)
+    fun action(input: String, session: EvaluationSession) {
+        // IMPLEMENT DEBUG BEHAVIOR HERE
+        val out = PrintStream(System.out)
+        val parser = PartiQLParserBuilder.standard().build()
+        val ast = parser.parseAstStatement(input)
+        val plan = AstToRel.convert(ast)
+        PlanPrinter.append(out, plan)
+    }
 }
