+++ partiql-cli/src/main/kotlin/org/partiql/cli/shell/Shell.kt
@@ -57,6 +57,18 @@ private const val PROMPT_2 = "   | "
 private const val BAR_1 = "===' "
 private const val BAR_2 = "--- "
 private const val WELCOME_MSG = "Welcome to the PartiQL shell!"
+private const val DEBUG_MSG = """    
+■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
+
+    ██████╗ ███████╗██████╗ ██╗   ██╗ ██████╗ 
+    ██╔══██╗██╔════╝██╔══██╗██║   ██║██╔════╝ 
+    ██║  ██║█████╗  ██████╔╝██║   ██║██║  ███╗
+    ██║  ██║██╔══╝  ██╔══██╗██║   ██║██║   ██║
+    ██████╔╝███████╗██████╔╝╚██████╔╝╚██████╔╝
+    ╚═════╝ ╚══════╝╚═════╝  ╚═════╝  ╚═════╝ 
+    
+■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
+"""
 
 private const val HELP = """
 !add_to_global_env  Adds to the global environment key/value pairs of the supplied struct
@@ -138,6 +150,11 @@ internal class Shell(
         out.info(WELCOME_MSG)
         out.info("Typing mode: ${compiler.options.typingMode.name}")
         out.info("Using version: ${retrievePartiQLVersionAndHash()}")
+        if (compiler is AbstractPipeline.PipelineDebug) {
+            out.println("\n\n")
+            out.success(DEBUG_MSG)
+            out.println("\n\n")
+        }
 
         while (!exiting.get()) {
             val line: String = try {
