--- partiql-lib/partiql-rel/src/main/kotlin/org/partiql/ir/debug/TreePrinter.kt
@@ -1,4 +1,0 @@
-package org.partiql.ir.debug
-
-
-// TODO
