+++ settings.gradle.kts
@@ -20,7 +20,7 @@ include(
     "partiql-types",
     "partiql-lib:partiql-isl",
     "partiql-lib:partiql-sprout",
-    "partiql-lib:partiql-rel",
+    "partiql-lib:partiql-plan",
     "test:partiql-tests-runner",
     "test:partiql-randomized-tests",
     "examples",
